Refactore Fetch Games With React Query
