+++ src/hooks/useGames.ts
@@ -1,6 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
 import type { GameQuery } from "../App";
-import useData from "./useData";
+
 import type { Platform } from "./usePlatform";
+
+import apiClient, { type FetchResponse } from "../services/api-client";
 
 export interface Game {
   id: number;
@@ -11,18 +14,22 @@ export interface Game {
   rating_top: number;
 }
 
-const useGames = (gameQuery: GameQuery) =>
-  useData<Game>(
-    "/games",
-    {
-      params: {
-        genres: gameQuery.genre?.id,
-        platforms: gameQuery.platform?.id,
-        ordering: gameQuery.sortOrder,
-        search: gameQuery.searchText,
-      },
+const useGames = (gameQuery: GameQuery) => {
+  return useQuery({
+    queryKey: ["games", gameQuery],
+    queryFn: () => {
+      return apiClient
+        .get<FetchResponse<Game>>("/games", {
+          params: {
+            genres: gameQuery.genre?.id,
+            parent_platforms: gameQuery.platform?.id,
+            ordering: gameQuery.sortOrder,
+            search: gameQuery.searchText,
+          },
+        })
+        .then((res) => res.data);
     },
-    [gameQuery]
-  );
+  });
+};
 
 export default useGames;
