+++ src/hooks/useGenre.ts
@@ -1,6 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import useData, { type FetchResponse } from "./useData";
-import apiClient from "../services/api-client";
+import apiClient, { type FetchResponse } from "../services/api-client";
 import genres from "../data/genres";
 export interface Genre {
   id: number;
