+++ src/hooks/usePlatform.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import platforms from "../data/platforms";
-import apiClient from "../services/api-client";
-import type { FetchResponse } from "./useData";
+import apiClient, { type FetchResponse } from "../services/api-client";
+
 export interface Platform {
   id: number;
   name: string;
