+++ src/services/api-client.ts
@@ -1,4 +1,8 @@
 import axios from "axios";
+export interface FetchResponse<T> {
+  count: number;
+  results: T[];
+}
 
 export default axios.create({
   baseURL: "https://api.rawg.io/api",
